% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mims.R
\name{mims}
\alias{mims}
\title{mims}
\usage{
mims(y, age, sex)
}
\arguments{
\item{y}{The activity level measurement}

\item{age}{The age of the individual (must be at least 3)}

\item{sex}{The sex of the individual (M for male and F for female)}
}
\value{
The percentile(s) of the provided activity measurements
}
\description{
Gets the activity percentile based on the NHANES (2003-2006) survey.
}
\examples{
mims(154754, 24, 'M')
}
